Create README.md
This example simulates the life of an ant colony. Each ant has its own position on the field and energy. Within 365 days, the actions of ants take place: movement, feeding and reproduction.

Ants move randomly within +/- 5 positions from their current position. If the new position is within the field (from 0 to 99), then the ant moves to it.

Ants eat every day, replenishing their energy by 3 units. If the energy reaches 10 units, the ant breeds, spending 10 units of energy, and a new ant is created in the colony.